Correções na story que subiu anteriormente
