+++ SME.GoogleClassroom.Testes.Unitario/Aplicacao/CasosDeUso/Alunos/Cursos/InserirAlunoCursoGoogleUseCaseTeste.cs
@@ -149,7 +149,7 @@ namespace SME.GoogleClassroom.Testes.Unitario.Aplicacao.CasosDeUso.Alunos.Cursos
             var cursoGoogle = new CursoGoogle("Curso Google", "Seção", 1010, 6, "criadorDoCurso.784512@edu.sme.prefeitura.sp.gov.br");
 
             mediator.Setup(a => a.Send(It.IsAny<ObterAlunosPorCodigosQuery>(), It.IsAny<CancellationToken>()))
-                .ReturnsAsync(new List<AlunoGoogle> { alunoGoogle });
+               .ReturnsAsync(new List<AlunoGoogle> { alunoGoogle });
 
             mediator.Setup(a => a.Send(It.IsAny<ObterCursoPorEmailNomeQuery>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(cursoGoogle);
@@ -213,7 +213,7 @@ namespace SME.GoogleClassroom.Testes.Unitario.Aplicacao.CasosDeUso.Alunos.Cursos
             var cursoGoogle = new CursoGoogle("Curso Google", "Seção", 1010, 6, "criadorDoCurso.784512@edu.sme.prefeitura.sp.gov.br");
 
             mediator.Setup(a => a.Send(It.IsAny<ObterAlunosPorCodigosQuery>(), It.IsAny<CancellationToken>()))
-                .ReturnsAsync(new List<AlunoGoogle> { alunoGoogle });
+               .ReturnsAsync(new List<AlunoGoogle> { alunoGoogle });
 
             mediator.Setup(a => a.Send(It.IsAny<ObterCursoPorEmailNomeQuery>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(cursoGoogle);
