+++ sme.gclass.application/Queries/Gsa/Cursos/ObterCursoGsaManualmente/ObterCursoGsaPorAnoQueryHandler.cs
@@ -2,21 +2,20 @@
 using SME.GoogleClassroom.Dados;
 using SME.GoogleClassroom.Infra;
 using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace SME.GoogleClassroom.Aplicacao
 {
-    public class ObterCursoGsaPorNomeQueryHandler : IRequestHandler<ObterCursoGsaPorNomeQuery, CursoGsaDto>
+    public class ObterCursoGsaPorAnoQueryHandler : IRequestHandler<ObterCursoGsaPorAnoQuery, IEnumerable<CursoGsaId>>
     {
         IRepositorioCursoGsa repositorioCursoGsa;
-        public ObterCursoGsaPorNomeQueryHandler(IRepositorioCursoGsa repositorioCursoGsa)
+        public ObterCursoGsaPorAnoQueryHandler(IRepositorioCursoGsa repositorioCursoGsa)
         {
             this.repositorioCursoGsa = repositorioCursoGsa ?? throw new ArgumentNullException(nameof(repositorioCursoGsa));
         }
-        public async Task<CursoGsaDto> Handle(ObterCursoGsaPorNomeQuery request, CancellationToken cancellationToken)
-        {
-            return await repositorioCursoGsa.ObterCursoGsaPorNomeAsync(request.Nome);
-        }
+        public Task<IEnumerable<CursoGsaId>> Handle(ObterCursoGsaPorAnoQuery request, CancellationToken cancellationToken)
+             => repositorioCursoGsa.ObterCursosGsaPorAno(request.AnoLetivo, request.CursoId, request.Pagina, request.QuantidadeRegistrosPagina);
     }
 }
